+++ .gitignore
@@ -170,5 +170,4 @@ cython_debug/
 # PyPI configuration file
 .pypirc
 
-BaseDeDados/
-BaseDeDados2/
+BasedeDados/
